node: drop depends from config to allow change versions on host
This change ensures that on unsupported targets like mpc85xx, you can
choose by using either ``make menuconfig`` or manually enable
CONFIG_NODEJS_version(14/16/18), which applies for node host version, you would like to use.

Before this change:
- I tried to use CONFIG_NODEJS_14 or CONFIG_NODEJS_16 for mvebu, where
  it works, but mpc85xx could not recognize it and fails with:
ERROR: Config option is not present in generated config: CONFIG_NODEJS_16=y

- So, always, on the unsupported target/router, where node is not
  available, it was using the latest node
  version, which is for all OpenWrt versions (currently) - 18.
It seems like some software is not ready for it, and it is a good idea
to expose version changing even for host.